requirements.txt
gradio~=5.27.0
sentence-transformers~=4.1.0
scikit-learn
numpy~=2.2.3
pandas~=2.2.3

tiktoken~=0.9.0
openai~=0.28.0
python-dotenv~=1.0.1
huggingface-hub~=0.30.2